Fix hashing of SCCs in dedupe
We hash an SCC by concatenating the hashes of all the nodes. Previously
we sorted the nodes by the hashes first, which meant two graphs with
identical nodes would get the same hash, even if the edges were
different. This commit changes the sorting to take topology into
account, thereby fixing the bug.